requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
boto3==1.13.11
botocore==1.16.11
click==7.1.2
cmake==3.17.2
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-storages==1.9.1
dlib==19.19.0
docutils==0.15.2
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
Pillow==7.1.2
psycopg2==2.7.7
pycodestyle==2.6.0
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.42
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
